+++ src/Command/CommandAbstract.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Pongee\DatabaseSchemaVisualization\Command;
 
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Connection\ConnectionCollection;
-use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Connection\NotDefinedConnection;
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\SchemaInterface;
 use Pongee\DatabaseSchemaVisualization\Parser\ParserInterface;
 use Symfony\Component\Console\Command\Command;
@@ -18,8 +17,6 @@ abstract class CommandAbstract extends Command
 {
     protected const string ARGUMENT_FILE = 'file';
 
-    protected const string OPTION_CONNECTION = 'connection';
-
     protected const string OPTION_TEMPLATE = 'template';
 
     protected string $rootDir;
@@ -31,38 +28,7 @@ abstract class CommandAbstract extends Command
 
         $this
             ->setHelpDescription()
-            ->addFileArgument()
-            ->addConnectionOption();
-    }
-
-    protected function addConnectionOption(): self
-    {
-        $this->addOption(
-            self::OPTION_CONNECTION,
-            null,
-            InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY,
-            "Force foreign keys definitions, if there are no foreign keys defined in sql schema.
-            Format: <info>child_table_name.child_table_columns=>parent_table_name.parent_table_columns</info>
-            --------
-            Example #1:
-            <info>foo.bar_id=>bar.id</info><comment>
-            It defines a foreign key definition between foo table's bar_id field and bar table's id field.
-            ALTER TABLE `foo` ADD FOREIGN KEY (`bar_id`) REFERENCES `bar` (`id`)
-            </comment>
-            --------
-            Example #2:
-            <info>foo.bar_id,type=>bar.id,type</info><comment>
-            It defines a foreign key definition between foo table's bar_id and type fields,
-            and bar table's id and type fields.
-            ALTER TABLE `foo` ADD FOREIGN KEY (`bar_id`, `type`) REFERENCES `bar` (`id`, `type`)</comment>
-            --------
-            Example #3:
-            <info>*.bar_id=>bar.id</info><comment>
-            It defines a foreign key definition between any table's bar_id field and bar table's id field.</comment>
-            "
-        );
-
-        return $this;
+            ->addFileArgument();
     }
 
     protected function addFileArgument(): self
@@ -92,52 +58,12 @@ abstract class CommandAbstract extends Command
     protected function setHelpDescription(): self
     {
         $this->setHelp(
-            sprintf(
-                "If the tables contain foreign keys it automatically resolves table connections.\n"
-                . "Otherwise the relations need to be defined under --%s parameter.\n"
-                . "The table connection types (one-to-one, one-to-many) are automatically resolved.\n"
-                . "For connection type resolving conditions please see REAMDME.",
-                self::OPTION_CONNECTION
-            )
+            "If the tables contain foreign keys it automatically resolves table connections.\n"
+            . "The table connection types (one-to-one, one-to-many) are automatically resolved.\n"
+            . "For connection type resolving conditions please see README."
         );
 
         return $this;
-    }
-
-    protected function getForcedConnections(array $connections): ConnectionCollection
-    {
-        $forcedConnectionCollection = new ConnectionCollection();
-        foreach ($connections as $connection) {
-            preg_match('#
-                ^
-                (?<childTableName>\S+)
-                \.
-                (?<childTableColumns>[^=]+)
-                =>
-                (?<parentTableName>[^.]+)
-                \.
-                (?<parentTableColumns>[^.]+)
-                $
-            #x', (string) $connection, $matches);
-
-            if (
-                !empty($matches['childTableName'])
-                && !empty($matches['childTableColumns'])
-                && !empty($matches['parentTableName'])
-                && !empty($matches['parentTableColumns'])
-            ) {
-                $forcedConnectionCollection->add(
-                    new NotDefinedConnection(
-                        $matches['childTableName'],
-                        $matches['parentTableName'],
-                        explode(',', $matches['childTableColumns']),
-                        explode(',', $matches['parentTableColumns'])
-                    )
-                );
-            }
-        }
-
-        return $forcedConnectionCollection;
     }
 
     protected function getSqlFileContent(InputInterface $input): string
@@ -174,7 +100,7 @@ abstract class CommandAbstract extends Command
     {
         return $this->parser->run(
             $this->getSqlFileContent($input),
-            $this->getForcedConnections($input->getOption(self::OPTION_CONNECTION))
+            new ConnectionCollection()
         );
     }
 }